Struct truth_tester::Tester

pub struct Tester<E: Expression> { /* fields omitted */ }

A struct used to store both the user given expression, and the State of that expression.

Methods

impl<E: ExprFn> Tester<E>

Linear implementation of all the Tester methods, based on an ExprFn.

pub fn passes(var_count: usize, expr: E) -> bool

Checks if any of possible states of the var_count variables fails the given expression.

This function returns true if all possible states pass the given expr. And false otherwise.

pub fn fails(var_count: usize, expr: E) -> bool

Checks if any of possible states of the var_count variables passes the given expression.

This function returns true if all possible states fail the given expr. And false otherwise.

pub fn succeeded(&self) -> bool

This returns true iff there are no failures

pub fn failed(&self) -> bool

This returns true iff there are no sucesses

pub fn successes<'a>(&'a self) -> impl Iterator<Item = State> + 'a

Iterate over all the successes in sequence

pub fn failures<'a>(&'a self) -> impl Iterator<Item = State> + 'a

Iterate over all the failures in sequence

pub fn eval<'a>(&'a self) -> impl Iterator<Item = (State, bool)> + 'a

Evaluate the expression of this Tester

impl<E> Tester<E> where
E: ExprFn + Send + Sync,

Parallel implementation of all the Tester methods, based on an ExprFn.

pub fn passes_par(var_count: usize, expr: E) -> bool

Checks if any of possible states of the var_count variables fails the given expression.

This function returns true if all possible states pass the given expr. And false otherwise.

This function is the parallel version of Tester::passes

pub fn fails_par(var_count: usize, expr: E) -> bool

Checks if any of possible states of the var_count variables passes the given expression.

This function returns true if all possible states fail the given expr. And false otherwise.

This function is the parallel version of Tester::fails

pub fn succeeded_par(&self) -> bool

This returns true iff there are no failures

This function is the parallel version of Tester::succeeded

pub fn failed_par(&self) -> bool

This returns true iff there are no sucesses

This function is the parallel version of Tester::failed

pub fn successes_par<'a>(&'a self) -> impl ParallelIterator<Item = State> + 'a

Iterate over all the successes in parallel

This function is the parallel version of Tester::successes

pub fn failures_par<'a>(&'a self) -> impl ParallelIterator<Item = State> + 'a

Iterate over all the failures in parallel

This function is the parallel version of Tester::failures

impl<E: ExprFn> Tester<E>

Tester based on an ExprFn.

pub fn new(var_count: usize, expr: E) -> Tester<E>

Create a new Tester with a given ExprFn.

impl<'t> Tester<Tokens<'t>>

Linear implementation of all the Tester methods, based on parsed Tokens.

pub fn passes<'i: 't>(inp: &'i str) -> bool

Checks if any of possible states of all variables in the given input expression fails.

This function returns true if all possible states pass the given inp, and false otherwise.

pub fn fails<'i: 't>(inp: &'i str) -> bool

Checks if any of possible states of all variables in the given input expression passes.

This function returns true if all possible states fail the given inp, and false otherwise.

pub fn succeeded(&self) -> bool

This returns true iff there are no failures

pub fn failed(&self) -> bool

This returns true iff there are no sucesses

pub fn successes<'b>(&'b self) -> impl Iterator<Item = State> + 'b

Iterate over all the successes in sequence

pub fn failures<'b>(&'b self) -> impl Iterator<Item = State> + 'b

Iterate over all the failures in sequence

pub fn eval<'b>(&'b self) -> impl Iterator<Item = (State, bool)> + 'b

Evaluate the expression of this Tester

impl<'t> Tester<Tokens<'t>>

Parallel implementation of all the Tester methods, based on parsed Tokens.

pub fn passes_par<'i: 't>(inp: &'i str) -> bool

Checks if any of possible states of all variables in the given input expression fails.

This function returns true if all possible states pass the given inp, and false otherwise.

This function is the parallel version of Tester::passes

pub fn fails_par<'i: 't>(inp: &'i str) -> bool

Checks if any of possible states of all variables in the given input expression passes.

This function returns true if all possible states fail the given inp, and false otherwise.

This function is the parallel version of Tester::fails

pub fn succeeded_par(&self) -> bool

This returns true iff there are no failures

This function is the parallel version of Tester::succeeded

pub fn failed_par(&self) -> bool

This returns true iff there are no sucesses

This function is the parallel version of Tester::failed

pub fn successes_par<'b>(&'b self) -> impl ParallelIterator<Item = State> + 'b

Iterate over all the successes in parallel

This function is the parallel version of Tester::successes

pub fn failures_par<'b>(&'b self) -> impl ParallelIterator<Item = State> + 'b

Iterate over all the failures in parallel

This function is the parallel version of Tester::failures

pub fn eval_par<'b>(
&'b self
) -> impl ParallelIterator<Item = (State, bool)> + 'b

Evaluate the expression of this Tester

This function is the parallel version of Tester::eval

impl<'t> Tester<Tokens<'t>>

Tester based on parsed Tokens.

pub fn parse<'i: 't>(inp: &'i str) -> Self

parse the given input into a Tester

pub fn parse_with_literals<'l, 'i: 't>(
inp: &'i str,
literals: TokenLiterals<'l>
) -> Self

parse the given input into a Tester using the given TokenLiterals

pub fn with_tokens(expr: Tokens<'t>) -> Self

Create a new Tester from the given Tokens